// src/types.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  MANAGER = 'MANAGER', // Usato globalmente o come ruolo specifico nel team
  USER = 'USER',
}

export interface Team {
  id: string;
  name: string;
  createdAt: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  role: UserRole; // Ruolo Globale (Admin vs Standard)
  
  // Multi-Team Support
  teamId?: string; // DEPRECATO (Mantenuto per retrocompatibilità temporanea)
  teamIds?: string[]; // Lista degli ID dei team a cui appartiene
  teamRoles?: Record<string, UserRole>; // Mappa { teamId: UserRole } per definire se è LEADER o MEMBER in quello specifico team

  isPending: boolean; 
  isDisabled?: boolean;
  photoURL?: string;
  isDeleted?: boolean;
}